+++ common/src/main/java/whocraft/tardis_refined/common/util/MiscHelper.java
@@ -30,6 +30,8 @@ import whocraft.tardis_refined.TardisRefined;
 import whocraft.tardis_refined.common.block.console.GlobalConsoleBlock;
 import whocraft.tardis_refined.common.block.life.EyeBlock;
 import whocraft.tardis_refined.common.block.shell.ShellBaseBlock;
+import whocraft.tardis_refined.common.block.shell.RedirectBlock;
+import whocraft.tardis_refined.common.blockentity.shell.ShellBaseBlockEntity;
 import whocraft.tardis_refined.common.capability.tardis.TardisLevelOperator;
 import whocraft.tardis_refined.common.protection.ProtectedZone;
 import whocraft.tardis_refined.registry.TRBlockRegistry;
@@ -188,7 +190,21 @@ public class MiscHelper {
             }
         }
 
-        return (state.getBlock() instanceof GlobalConsoleBlock && world.dimensionTypeId() == TRDimensionTypes.TARDIS) || state.getBlock() instanceof ShellBaseBlock || state.getBlock() instanceof EyeBlock;
+        if (state.getBlock() instanceof RedirectBlock redirect) {
+            var foundPos = redirect.findSource(world, pos, state);
+            if (foundPos.isPresent()) {
+                var foundState = world.getBlockState(foundPos.get());
+                if (!(foundState.getBlock() instanceof RedirectBlock)) {
+                    return shouldCancelBreaking(world, entity, foundPos.get(), foundState);
+                }
+            }
+        }
+
+        if (state.getBlock() instanceof ShellBaseBlock && world.getBlockEntity(pos) instanceof ShellBaseBlockEntity shell) {
+            return shell.getTardisId() != null;
+        }
+
+        return (state.getBlock() instanceof GlobalConsoleBlock && world.dimensionTypeId() == TRDimensionTypes.TARDIS) || state.getBlock() instanceof EyeBlock;
     }
 
     public static Component getTranslatableDimensionName(ResourceKey<Level> dimensionKey) {