+++ changelog.md
@@ -12,6 +12,9 @@
 - Pressing E will now close Monitor, ARS, and Astral Manipulator GUIs.
 - Improved keyboard navigation support for the Monitor and Astral Manipulator GUIs.
 - The TARDIS item will now read the `setup_data` tag on the item, if it exists, and pass it on to the created TARDIS.
+- Shells are now breakable if no TARDIS dimension exists yet. This means the root shell is now breakable before you shear it open.
+- Shulker shell now uses a 1 block tall bounding box when closed.
+- Addons can now specify arbitrary voxel shapes for their shell and interior door collision boxes.
 
 #### Bug Fixes
 
@@ -30,3 +33,8 @@
 - Fixed items dropped in shell view ending up outside the TARDIS.
 - Fixed players being able to pick up items and XP while in shell view.
 - Fixed the TARDIS sometimes playing the default hum when set to a different hum.
+- Fix root shell not breaking when placed inside a TARDIS.
+- Fix top half of door and shell bounding box not being selectable. A hidden redirect block will now be placed there which you can see in F3 when sneaking. Don't worry, the block is only placed if there's space, so the diamond block you glitched into the top half of your TARDIS shell is still safe.
+- Fix briefcase shell and door using the default bounding box.
+- Fix Immersive Portals portal not positioned correctly for offset interior doors.
+- Fix half-baked shell no longer being shearable after taking off and landing (or reloading the world). It will now continue to be shearable until you unlock the chameleon circuit.